Move BaseConstructors to the platform.base package
