+++ src/main/kotlin/org/powernukkit/converters/platform/bedrock/entity/BedrockEntity.kt
@@ -19,7 +19,7 @@
 package org.powernukkit.converters.platform.bedrock.entity
 
 import org.powernukkit.converters.math.EntityPos
-import org.powernukkit.converters.platform.base.block.BaseConstructors
+import org.powernukkit.converters.platform.base.BaseConstructors
 import org.powernukkit.converters.platform.base.entity.BaseEntity
 import org.powernukkit.converters.platform.bedrock.BedrockPlatform
